Reduce duplication between publish and unpublish commands
